Fix issue when iOS is not properly placing SVG
Issue related to iOS not well handling foreignObject with ha-icon (known HA issue since 2020). Solution proposed is to use the path instead and for this it is needed to embed the mdi/js, which is part of HA anyhow.
For other icon-groups/families like simpleicons, it needs to be reviewed if there is a serious demand as handling those is more complex as SVG or (alt.) less complex as image....t.b.d. for later